refactor: InpuBar as subview instead of accessoryView
Move typing indicator methods to extension
